test: split up annotations twest in twain
